refactor(frontend): group provenance UI into provenance/
Move the claim, attribution and qualifier display components into a
provenance/ folder and update importers. Add a README marking the folder
read-only: claim editing happens through the section editors.

This adds the first markdown file under frontend/, whose .prettierrc loads
prettier-plugin-svelte. Format frontend markdown from the frontend/ working
directory in pre-commit so the plugin resolves.
