+++ frontend/src/lib/components/provenance/AttributionLine.svelte
