+++ frontend/src/lib/components/provenance/ClaimValue.fixture.svelte
