+++ frontend/src/lib/components/provenance/README.md
@@ -1,0 +1,8 @@
+# provenance
+
+This folder contains provenance display components — claim values, attribution,
+authorship and qualifiers.
+
+These components are only for DISPLAY of provenance information; editing of
+claim values happens through the section editors, not here. Form components
+do not belong in this folder.
