+++ frontend/src/routes/display-types/[slug]/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import client from '$lib/api/client';
-  import AttributionLine from '$lib/components/AttributionLine.svelte';
+  import AttributionLine from '$lib/components/provenance/AttributionLine.svelte';
   import Markdown from '$lib/components/Markdown.svelte';
   import TitleCard from '$lib/components/cards/TitleCard.svelte';
   import PaginatedSection from '$lib/components/grid/PaginatedSection.svelte';
