+++ frontend/src/routes/reward-types/[slug]/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import AttributionLine from '$lib/components/AttributionLine.svelte';
+  import AttributionLine from '$lib/components/provenance/AttributionLine.svelte';
   import Markdown from '$lib/components/Markdown.svelte';
   import ClientFilteredGrid from '$lib/components/grid/ClientFilteredGrid.svelte';
   import MachineCard from '$lib/components/cards/MachineCard.svelte';
