+++ frontend/src/routes/tags/[slug]/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import client from '$lib/api/client';
-  import AttributionLine from '$lib/components/AttributionLine.svelte';
+  import AttributionLine from '$lib/components/provenance/AttributionLine.svelte';
   import Markdown from '$lib/components/Markdown.svelte';
   import MachineCard from '$lib/components/cards/MachineCard.svelte';
   import PaginatedSection from '$lib/components/grid/PaginatedSection.svelte';
